feat: 작업 편집 시 담당자 이메일 필드 추가
- 간트차트 및 칸반보드 작업 편집 폼에 이메일 입력 추가
- 빈 값은 null로 전송 처리
- 간트차트 작업 삭제 완료 알림 추가
